Optimize chunk reader
Ok, so really didn't intend to do this but was trying to work out an
issue I (thought) I was having which turned out to not be an issue, but
in doing so re-arranged the reader to not have to copy buffers around. A
little messier, maybe not worth it, but already done.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>